rename some js to match webapp template filenames for easier searching
