+++ _front-end/src/assets/toolkit/scripts/global.js
@@ -10,8 +10,8 @@ import revealToggle from './components/reveal-toggle';
 import invalidApiKey from './components/invalidApiKey';
 import login from './sections/login';
 import home from './sections/home';
-import vocab from './sections/vocab';
-import unlocks from './sections/unlocks';
+import vocab from './sections/vocabulary';
+import levelVocab from './sections/levelVocab';
 import reviews from './sections/reviews';
 
 $(document).ready(() => {
@@ -22,7 +22,7 @@ $(document).ready(() => {
   login.init();
   home.init();
   vocab.init();
-  unlocks.init();
+  levelVocab.init();
   reviews.init();
 
 })
