+++ _front-end/src/assets/toolkit/scripts/sections/vocabulary.js
